+++ src/main/kotlin/io/github/facilityapi/intellij/FsdUsageTypeProvider.kt
@@ -1,0 +1,9 @@
+package io.github.facilityapi.intellij
+
+import com.intellij.psi.PsiElement
+import com.intellij.usages.impl.rules.UsageType
+import com.intellij.usages.impl.rules.UsageTypeProvider
+
+class FsdUsageTypeProvider : UsageTypeProvider {
+    override fun getUsageType(element: PsiElement?): UsageType = UsageType.CLASS_FIELD_DECLARATION
+}