+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ## [Unreleased]
 ### Added
 - Rename types with inline refactoring
+- Autocomplete user-defined types
+- Find usages for user-defined types
 - Comment or uncomment lines with editor actions
 
 ## 0.0.4